Вынесение карточек из "Избранное" в отдельные виджеты
