+++ lib/ui/screen/sight_card.dart
@@ -5,7 +5,7 @@ import 'package:places/ui/screen/sight_details.dart';
 
 import '../../globals.dart';
 
-class SightCard extends StatelessWidget {
+class SightCard extends StatefulWidget {
   const SightCard({
     key,
     required this.sight,
@@ -14,113 +14,9 @@ class SightCard extends StatelessWidget {
   final Sight sight;
 
   @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        Container(
-          decoration: BoxDecoration(
-              borderRadius: BorderRadius.only(
-                  topLeft: Radius.circular(12), topRight: Radius.circular(12))),
-          child: Column(
-            children: [
-              Container(
-                height: 96,
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.only(
-                      topLeft: Radius.circular(12),
-                      topRight: Radius.circular(12)),
-                  gradient: new LinearGradient(
-                      colors: [
-                        Colors.black38,
-                        Colors.black12,
-                      ],
-                      begin: const FractionalOffset(0.0, 0.0),
-                      end: const FractionalOffset(1.0, 0.0),
-                      stops: [0.0, 1.0],
-                      tileMode: TileMode.clamp),
-                ),
-                child: Stack(
-                  fit: StackFit.expand,
-                  children: [
-                    ClipRRect(
-                      borderRadius: BorderRadius.only(
-                          topLeft: Radius.circular(12),
-                          topRight: Radius.circular(12)),
-                      child: Image.network(
-                        sight.url,
-                        loadingBuilder: (BuildContext context, Widget child,
-                            ImageChunkEvent? loadingProgress) {
-                          if (loadingProgress == null) {
-                            return child;
-                          }
-                          return Center(
-                            child: CircularProgressIndicator(
-                              value: loadingProgress.expectedTotalBytes != null
-                                  ? loadingProgress.cumulativeBytesLoaded /
-                                      loadingProgress.expectedTotalBytes!
-                                  : null,
-                            ),
-                          );
-                        },
-                        fit: BoxFit.fitWidth,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              Container(
-                decoration: BoxDecoration(
-                    borderRadius: BorderRadiusDirectional.only(
-                      bottomEnd: Radius.circular(12),
-                      bottomStart: Radius.circular(12),
-                    ),
-                    color: isDarkMode
-                        ? dmSightCardContainerColor
-                        : lmSightCardContainerColor),
-                alignment: Alignment.centerLeft,
-                padding: EdgeInsets.all(16.0),
-                child: informationColumn(context),
-              ),
-            ],
-          ),
-        ),
-        new Positioned.fill(
-            child: new Material(
-                color: Colors.transparent,
-                child: new InkWell(
-                  onTap: () {
-                    Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => SightDetails(
-                                  sight: sight,
-                                )));
-                  },
-                  customBorder: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.all(Radius.circular(12)),
-                  ),
-                ))),
-        Container(
-          margin: EdgeInsets.only(left: 16),
-          child: Row(
-            children: [
-              Expanded(
-                child: Container(
-                  child: Text(
-                      sight.type.isEmpty
-                          ? "категория"
-                          : sight.type.toLowerCase(),
-                      style: Theme.of(context).primaryTextTheme.caption),
-                ),
-              ),
-              interactionButtons(),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
+  _SightCardState createState() => _SightCardState();
 
+  
   Widget interactionButtons() {
     return Container(
       width: 45,
@@ -156,4 +52,115 @@ class SightCard extends StatelessWidget {
       ),
     ]);
   }
+}
+
+class _SightCardState extends State<SightCard> {
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      children: [
+        Container(
+          decoration: BoxDecoration(
+              borderRadius: BorderRadius.only(
+                  topLeft: Radius.circular(12), topRight: Radius.circular(12))),
+          child: Column(
+            children: [
+              Container(
+                height: 96,
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.only(
+                      topLeft: Radius.circular(12),
+                      topRight: Radius.circular(12)),
+                  gradient: new LinearGradient(
+                      colors: [
+                        Colors.black38,
+                        Colors.black12,
+                      ],
+                      begin: const FractionalOffset(0.0, 0.0),
+                      end: const FractionalOffset(1.0, 0.0),
+                      stops: [0.0, 1.0],
+                      tileMode: TileMode.clamp),
+                ),
+                child: Stack(
+                  fit: StackFit.expand,
+                  children: [
+                    ClipRRect(
+                      borderRadius: BorderRadius.only(
+                          topLeft: Radius.circular(12),
+                          topRight: Radius.circular(12)),
+                      child: Image.network(
+                        widget.sight.url,
+                        loadingBuilder: (BuildContext context, Widget child,
+                            ImageChunkEvent? loadingProgress) {
+                          if (loadingProgress == null) {
+                            return child;
+                          }
+                          return Center(
+                            child: CircularProgressIndicator(
+                              value: loadingProgress.expectedTotalBytes != null
+                                  ? loadingProgress.cumulativeBytesLoaded /
+                                      loadingProgress.expectedTotalBytes!
+                                  : null,
+                            ),
+                          );
+                        },
+                        fit: BoxFit.fitWidth,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              Container(
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadiusDirectional.only(
+                      bottomEnd: Radius.circular(12),
+                      bottomStart: Radius.circular(12),
+                    ),
+                    color: isDarkMode
+                        ? dmSightCardContainerColor
+                        : lmSightCardContainerColor),
+                alignment: Alignment.centerLeft,
+                padding: EdgeInsets.all(16.0),
+                child: widget.informationColumn(context),
+              ),
+            ],
+          ),
+        ),
+        new Positioned.fill(
+            child: new Material(
+                color: Colors.transparent,
+                child: new InkWell(
+                  onTap: () {
+                    Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) => SightDetails(
+                                  sight: widget.sight,
+                                )));
+                  },
+                  customBorder: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(12)),
+                  ),
+                ))),
+        Container(
+          margin: EdgeInsets.only(left: 16),
+          child: Row(
+            children: [
+              Expanded(
+                child: Container(
+                  child: Text(
+                      widget.sight.type.isEmpty
+                          ? "категория"
+                          : widget.sight.type.toLowerCase(),
+                      style: Theme.of(context).primaryTextTheme.caption),
+                ),
+              ),
+              widget.interactionButtons(),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+
 }
